Clean CLion project, Do visual testing on algo
The CLion project files were incorrectly formatted in the .gitignore,
so a bunch of unnecessary files were added to github. This has been
fixed.

The algorithm was modified such that I could do some basic testing
to visualise where to cuboids were placed in the box. It is not
yet complete though.